This commit defines the structure of the virtual machine for the MIR interpreter. I think it's important for this to be provided on its own, as it will make it easier to understand how the evaluation functions which step through the MIR behave. (These will be provided incrementally in upcoming PRs.). The code is heavily commented to explain why certain structures contain the information they do.
The memory model is a rudimentary byte array, represented as a vector with each
item initialized to 0. The max size can be provided upon construction but I
haven't decided on what an appropriate value for this is yet.

One important aspect of the Machine's memory is that it is aligned according to
the min. required by the ABI. Allowing me to allocate ptrs guaranteed to be
correctly aligned.

    *See std::mem::align_of.

It is possible this may not be the correct alignment for my needs, the docs
state this function could return an alignment smaller than the one preferred
for the target architecture. I haven't thought too deeply about the implications
of this yet.

Memory cannot yet be deallocated, this will be implemented in the future.