Included basic interpreter methods for evaluating MIR constructs.
Currently, only assignment is implemented for the simplicity of review.
Upcoming commits will unstub match cases with their respective
functionality.

In addition, some slight changes have been made to the Machine.
Primarily, the frame no longer owns the MIR, but instead holds an
immutable reference to it, whose lifetime is bound by the type context.
This is an important change, a Frame represents an invocation of a
function, and the MIR exists for each declaration. This would have
resulted in the need for copying or hairy lifetime management.