Implemented compiler hook to grab MIR when given path to rust program.
When given a path to a valid rust program as a CLI, Mirkat can halt
compilation and access Rust's MIR. This gives us an entry point to pass
the MIR to the interpreter and begin running it in the Mirkat virtual
machine.

Rust complains if you don't specify a sysroot when using the unstable
rustc crate. So I've borrowed code from Miri which looks for an
environment variable MIRKAT_SYSROOT before passing it as a CLI to the
compiler.