Interpreter methods now implemented on Machine struct directly
It's very common for these methods to mutate the Machine struct
directly, at least, they always require a mutable reference to the
machine. Passing a mutable reference to the vm as the final parameter
felt like a bit of an anti-pattern, at the very least it increased
unnecessary code bloat with method signatures.

This has also simplified a few of the dances I had to perform to get
around the borrow checker.

Less importantly, this is also convenient in that it allows us to define
new interpret methods which can be bound implicitly to the lifetimes of
the machine struct. The code is a little more readable this way.