Replaced HashMap storage of local vars on stackframes with IndexedVec.
HashMaps are inefficient for storing locals. They use up more space than
necessary and the hash function is non-deterministic. Statically, the
MIR gives us unique integer IDs for Locals. These can be directly mapped
to indices into the frame's local storage vector.

Technically, this collection is an IndexVec, a rustc datastructure
which acts as a wrapper around a vec. This provides static type checking
on the indices for sanity. The Local type is simply a tuple struct with
a single u32 element, but encoding it this way means that it is
considered it's own type. Other integer identifiers exist throughout
rustc and are encoded the same way (e.g BasicBlocks). Using this
tuple struct type ensures that, for example, we can never accidently use a
BasicBlock to refer to a Local's index.

Local's values are stored as vectors of bytes, these are optional so
that they can be lazily instantiated for the sake of optimisation.