Implemented interpretation of the Rust bottom type `()`.
The bottom type exists in every Rust program at least once - it is the
return value of the main function. Therefore, in order to interpret even
the most trivial of Rust programs, the interpreter has some way of
dealing with an empty tuple definition.

Currently, when Mirkat encounters this, it will do a noop and move on.
The empty tuple should use no memory, so this seems appropriate for now.