Implemented tuple de/serialization to byte vectors
In this approach, aggregate Values store boxed references to other
Values. Each tuple item doesn't need to store type information as this
is available from the tuple type. This requires the to_bytes function to
be moved. It is now implemented on the Value enum and called from the
TyVal container.