Improved locals to be exact sized and better field access.
The space required for frame local variables is known to us statically
from the MIR struct. We now take advantage this by initializing
local variables to zeroed byte arrays of the exact size needed for their
given type. This has the following benefits:

    - It simplifies the code required to read and write to local vars.
    There are less unnecessary nested option types.

    - It should* be more performant. When storing aggregate types, Rust
    does this on a field by field basis. Byte vectors used for storage
    may have previously needed to be dynamically resized in order to
    store additional field data. Worse still, if fields were not
    inserted in sequential order, inserts would be very slow as each
    byte after would need shifting to the right.

    - Frames are now more space efficient. Each local only has allocated
    the exact amount of memory it will ever need.

Rust's MIR uses `Projection`s to get and set fields of aggregate types.
A Projection can be considered a field access or a pointer chase to a
position in memory.

In order to accomodate for this in the Mirkat memory model, `Address`
enum variants for local variables now contain an optional sized pointer
struct for accessing only data in a specific field. This is always
checked upon access to prevent against a buffer overflow.

* This is speculative, I haven't got benchmarks to back this up