Use fixed size byte vector for frame local storage.
The motivation for this is to improve performance for frame creation and
local variable lookups. Upon frame creation, we create a statically
sized contiguous byte vector. This should reduce the chance that locals
are fragmented in memory and improve cache locality.

There is a slight overhead on local variable lookups, as a pointer for
the appropriate offset into the byte vector needs to be retrieved upon
access time. These are calculated upon frame creation and cached in the
`idxs` vec field on the Frame struct. This will allow us to use the
`Local` index in the MIR to obtain the correct offset into the local
storage vector.