Implemented support for checked binary addition and equality.
In Rust, checked binary addition on two integers results in a tuple:
(i_type, bool)

Where the first index is the result and the second records whether an
overflow happened.

This is important for the resulting control flow - a checked binary
operation will always be the last statement in a given basic block. If
it overflows, or some other exception happens, Rust will panic. This is
written into the control flow as an assert terminator.