Moved redundant Frame fields into Function struct.
A Frame struct represents a single function call. These Frames make use
of common metadata which is the same for every call of function F. For
this reason, we extract this common metadata and store it in a longer
living Function struct:

    struct Frame {
        function: Rc<Function<'tcx>>
    }

    struct Function {
        ... // metadata
    }

This is more memory efficient but requires runtime reference counting in
order to work. A benchmark will be forthcoming to measure the
performance change of this commit.